defer recalc operations; moved scrolling methods to own file
