+++ src/lib/modules.js
@@ -1,0 +1,5 @@
+'use strict';
+
+module.exports = {
+    Scrollbar: require('finbars')
+};
